WEBAPP-126: Move LocationLayout to layout module
