+++ src/modules/app/containers/RouterFragment.js
@@ -4,7 +4,7 @@ import React from 'react'
 import Layout from 'modules/layout/components/Layout'
 import GeneralHeader from '../../layout/components/GeneralHeader'
 import GeneralFooter from '../../layout/components/GeneralFooter'
-import LocationLayout from './LocationLayout'
+import LocationLayout from '../../layout/containers/LocationLayout'
 
 import SearchPage from 'routes/search/containers/SearchPage'
 import DisclaimerPage from 'routes/disclaimer/containers/DisclaimerPage'
